Add new array front matter called notaxonomy
Since the last update Vitae shows every taxonomy a post is part of below the
content. This could lead to a very unpleasantly large number of taxonomies
under the post.

With the new front matter notaxanomy you can specify every taxonomy that
should be not visible below the post.

Like this -> notaxonomy = ["series","tags","authors"]