Update content links with relURL feature
In the past there where problems with the use of Vitae in a subfolder of
a domain. For example "example.com/blog/". This update solves this
problem with the use of relURL, a feature that is implemented in Hugo.

*ATTENTION*
If you use the custom.css feature be aware that you need to change the
path now. Before the update, the custom.css file was located in
"static/css/" by default. If you still want to use this path just change
the path to "css/custom.css" in your config file. For an example see the
config file in "exampleSite/".