Add cmd option for .Site.Params.social
The url option is as the name suggests only for urls. If the social icon
should use a commands like "mailto:example@example.org" the new option
cmd should be used instead of url.

This is a reaction of the commit who updated links to use relURL. This
update caused problems with commands like "mailto:" this is now the new
way to go.

For an example see the config file in exampleSite/