Fix trailing whitespace in rendered link
Rendered links that are created with a own version of `render-link.html`
add a newline if the code is not in a single line and the end of the
file has a trailing whitespace.

This is a fix for this issue.

The Hugo maintainers know about the problem and have passed it on to Go
upstream.