Fix colors for a better look and feel in dark mode
Some of the former colors were not suitable for dark mode and had to be
replaced.

Be aware that this might change the look and feel of your website also
in normal mode. Check the colors before applying them to your live
website.

If you want to change the colors use the customCSS feature of Vitae.